% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{dentrassi}
\alias{dentrassi}
\title{Function to create a random sample of Dentrassi}
\usage{
dentrassi(n, stats = stats_index(), ...)
}
\arguments{
\item{n}{numeric. Number of individuals to create.}

\item{stats}{which stats to include in the sample. See 'stats_index()' for a
list of available stats.}

\item{...}{currently serves no function.}
}
\description{
Function to create a random sample of Dentrassi
}
\details{
In the current edition of The Guide, there is no
entry for the Dentrassi. However, they are described as 'an unruly tribe of
gourmands, a wild but pleasant bunch whom the Vogons had recently taken to employ
as catering staff on their long-haul fleets (...)'.

-- \emph{Douglas Adams, Hitchiker's Guide to the Galaxy}
}
