% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\docType{data}
\name{dentrassi_occupations}
\alias{dentrassi_occupations}
\title{Dentrassi Occupations}
\format{
A data frame with ? rows and 1 variable.
}
\source{
include source...
}
\usage{
dentrassi_occupations
}
\description{
A list of dentrassi occupations
}
\keyword{datasets}
