% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h2g2}
\alias{h2g2}
\title{Create a random sample}
\usage{
h2g2(n, race = race_index(), na_prob = NULL, ...)
}
\arguments{
\item{n}{number of individuals to generate.}

\item{race}{which races to include in the generated sample. Default is all.}

\item{...}{can be used to pass 'stats' through to race-specific functions. see example.}

\item{missing_prob}{Numeric. Proportion of values that should be missing at random.
Must be a number between 0 and 1.}
}
\description{
Create a random sample
}
