% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\docType{data}
\name{human_occupations}
\alias{human_occupations}
\title{Human Occupations}
\format{
A data frame with ? rows and 1 variable.
}
\source{
include source...
}
\usage{
human_occupations
}
\description{
A list of human occupations
}
\keyword{datasets}
