% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{humans}
\alias{humans}
\title{Random sample of humans}
\usage{
humans(n, stats = stats_index(), ...)
}
\arguments{
\item{n}{numeric. Number of individuals to create.}

\item{stats}{which stats to include in the sample. See 'stats_index()' for a
list of available stats.}

\item{...}{currently serves no function.}
}
\description{
The entry in The Guide for Earth: "Mostly harmless."
}
\details{
Function to create a random sample of humans
}
