% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iidr}
\alias{iidr}
\title{Same functionality as the 'infinite_improbability_drive' function but infinitely
easier to write!}
\usage{
iidr(...)
}
\arguments{
\item{...}{can be used to pass 'stats' through to race-specific functions. see example.}
}
\description{
Same functionality as the 'infinite_improbability_drive' function but infinitely
easier to write!
}
