% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inf_improb_dr}
\alias{inf_improb_dr}
\title{Generates a random statement indicating the odds of an event}
\usage{
inf_improb_dr()
}
\description{
Generates a random statement indicating the odds of an event
}
