% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{infinite_improbability_drive}
\alias{infinite_improbability_drive}
\title{Generates a random statement indicating the odds of an event}
\usage{
infinite_improbability_drive()
}
\description{
Generates a random statement indicating the odds of an event
}
