% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing}
\alias{missing}
\title{Inserts NAs throughout a data frame at random}
\usage{
missing(x, na_prob = 0.1, cols = seq_along(x), symm = F)
}
\arguments{
\item{x}{Data frame or vector.}

\item{na_prob}{the probability of missing values per variable in the data frame.
Default is 0.1, i.e. each variable will have 10\% NAs.}

\item{cols}{numeric. In which columns to insert NAs. Default is all columns.
To leave out the first column, use cols = -1.}

\item{symm}{logical. whether the missing values should be inserted symmetrically,
i.e., whether each column should contain the same number of NAs.}
}
\value{
Returns the original data frame with NAs inserted. The default settings
returns a data frame with data missing completely at random.
}
\description{
Inserts NAs throughout a data frame at random
}
\details{
Function is a modified version of 'r_na' and r_na_vector' from the
'wakefield' package. It adds the functionality to add NAs asymmetrically to
a data frame, meaning the number of NAs per column will vary.
}
\seealso{
\code{\link[wakefield]{r_na}}
}
