% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{towel}
\alias{towel}
\title{This function is about the most massively useful thing in this package.}
\usage{
towel(x, rowwise = F, colwise = T, cols = NULL)
}
\description{
This function is about the most massively useful thing in this package.
}
