% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations.R
\docType{data}
\name{vogon_occupations}
\alias{vogon_occupations}
\title{Vogon Occupations}
\format{
A data frame with ? rows and 1 variable.
}
\source{
same as humans...
}
\usage{
vogon_occupations
}
\description{
A list of vogon occupations. The original list was the same as the human
occupations, but with all non-administrative jobs removed.
}
\keyword{datasets}
