% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{golgafrinchans}
\alias{golgafrinchans}
\title{Function to create a random sample of Golgafrinchans}
\usage{
golgafrinchans(n, stats = stats_index(), ...)
}
\arguments{
\item{n}{numeric. Number of individuals to create.}

\item{stats}{which stats to include in the sample. See 'stats_index()' for a
list of available stats.}

\item{...}{currently serves no function.}
}
\description{
Function to create a random sample of Golgafrinchans
}
