% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing}
\alias{missing}
\title{Inserts NAs throughout a data frame at random}
\usage{
missing(x, na_prob = 0.1, cols = seq_along(x))
}
\arguments{
\item{x}{Data frame or vector.}

\item{na_prob}{the probability of missing values per variable in the data frame.
Default is 0.1, i.e. each variable will have 10\% NAs.}

\item{cols}{numeric. In which columns to insert NAs. Default is all columns.
To leave out the first column, use cols = -1.}
}
\description{
Inserts NAs throughout a data frame at random
}
\details{
Function is a modified version of 'r_na' and r_na_vector' from the
'wakefield' package.
}
\seealso{
\code{\link[wakefield]{r_na}}
}
