% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vogon_poetry}
\alias{vogon_poetry}
\title{This function creates randomly generated Vogon poetry, thought to be the third
worst poetry in the universe, only surpassed by that of the Azgoths of Kria,
and Paula Nancy Millstone Jennings of Sussex.}
\usage{
vogon_poetry()
}
\description{
This function creates randomly generated Vogon poetry, thought to be the third
worst poetry in the universe, only surpassed by that of the Azgoths of Kria,
and Paula Nancy Millstone Jennings of Sussex.
}
